Stub code for implementing the <include> <import> resolution in run-time. The treeTable instances are defined per XML schema file so there might be more than one  treeTable instance with the same target namespace.
NOTE: In the case of <include> - both the host- and included-schema must have identical target namespace.
That is, the current implementation does not support the case of using <include> from XML Schema
that has non empty target namespace and the referenced/included XML schema have an empty target
namespace, although this is allowed by the XML Schema spec.

git-svn-id: svn://svn.code.sf.net/p/exip/code/trunk@298 3073ab62-7c58-44d6-84d6-47b9a845c6e8